App/RoomsScreen - + extracted rooms flat list to it's own component RoomsListView
